Splitting job of making help from formatting help text.
